Config: prioritize cachepath from config file over system's XDG_HOME_CACHE
Remove redondant Path() and change cachepath resolution order.

Reasoning:
XDG_HOME_CACHE or HOME/.cache is expected to be set by default in the OS.
Thus, any value given to cachepath in the config file was to be overridden.

Inverting the priority insure that any given cachepath value will be used,
unless it is overriden by the command line option.

Signed-off-by: Alexandre Demers <alexandre.f.demers@gmail.com>